fix: add long openapi v3 schema file test
The long ingresscontroller openapi v3 schema is truncated if saved to a
bash variable and printed to stdout. As a result, we do not save long
text in bash variables and instead pipe them directly through functions.

To conclude, for long files you need to always write to a file.