Build Docker Image (#7)
* refactor: inline version into script

* refactor: package crd2jsonschema with nix

* refactor: split long kcov command into multiple lines

* fix: nix shebang shenanigans

I used `#!/usr/bin/env node` and `#!/usr/bin/env bash` shebangs which
are portable across most linux distributions which adhere to the new
[/usr
merge](https://www.freedesktop.org/wiki/Software/systemd/TheCaseForTheUsrMerge/).

Nix, however, does not adhere to this. As a result, we need to patch
every shebang with the `patchShebangs` function. Otherwise binaries
are not found and lead to script/test failures.

* feat: build docker image with nix

* refactor: create docker integration test

* feat: add functions to deploy docker image

* feat: deploy multi arch image to ghcr.io registry

Besides, it adds the latest tag for the latest version.

* feat: print help test if called without arguments

* fix: all.json should only be written with -a option

The all.json files was always written due to a bash typo of
`-:` instead of correct `:-`. Since I check for non-zero length
I decided to use the `-` only because it is equivalent to `:-` for
this case and prevents such typos in the future.

* refactor: group similar tests together

* refactor: harmonize output error message

* refactor: call test with main path

* refactor: add examples to usage message

* refactor: extend docker integration test

* refactor: add multiple files docker integration test

* chore: update readme

* chore: use permanent code link

* chore: add licence to readme and inline usage code