refactor: resolve breaking changes of dependency
The openapi-schema-to-json-schema update from v3 -> v5 brought some
breaking changes. This commit integrates the new v5 API.

Besides, it puts the built dist/oas3tojsonschema4 binary in front of the
path because otherwise it would take an old version of the
oas3tojsonschema4 binary which might be installed on the system which is
wrong because it does not contain the changes to test and might lead to
strange behaviour: e.g. the tests work locally but would fail in the CI
because the CI does not have the old oas3tojsonschema4 binary installed.
As a result, we can again run the tests locally with `bats test` and be
confident that we test the current changes and not and old version of
the oas3tojsonschema4 binary.