fix: remove the oas3tojsonschema4 binary from bin
because it is not intended to be used by the end user and every binary
in the `$out/bin` directory will be appended to the path. Instead put
the binary into the `$out/libexec` folder which is meant for binaries
which are called by the package but should not be called by the user
directly.