Update openapi-schema-to-json-schema 3.2.0 -> 5.1.0 (#9)
* chore: update flake

* chore: update openapi-schema-to-json-schema 3.2.0 -> 5.1.0

* chore: update crd2jsonschema version 1.0.0 -> 1.0.1

* refactor: resolve breaking changes of dependency

The openapi-schema-to-json-schema update from v3 -> v5 brought some
breaking changes. This commit integrates the new v5 API.

Besides, it puts the built dist/oas3tojsonschema4 binary in front of the
path because otherwise it would take an old version of the
oas3tojsonschema4 binary which might be installed on the system which is
wrong because it does not contain the changes to test and might lead to
strange behaviour: e.g. the tests work locally but would fail in the CI
because the CI does not have the old oas3tojsonschema4 binary installed.
As a result, we can again run the tests locally with `bats test` and be
confident that we test the current changes and not and old version of
the oas3tojsonschema4 binary.

* fix: remove the oas3tojsonschema4 binary from bin

because it is not intended to be used by the end user and every binary
in the `$out/bin` directory will be appended to the path. Instead put
the binary into the `$out/libexec` folder which is meant for binaries
which are called by the package but should not be called by the user
directly.

* refactor: remove shellcheck -x option

The crd2jsonschema.sh file does not use external sources. As a result,
the -x option can be removed.